+++ shared/src/db/speed_times.ts
@@ -14,7 +14,7 @@ export async function createSpeedTimesTable(db: knex): Promise<void> {
   const hasTable = await db.schema.hasTable(SPEED_TIMES_TABLE_NAME);
   if (!hasTable) {
     await db.schema.createTable(SPEED_TIMES_TABLE_NAME, table => {
-      table.integer(SpeedTimesColumn.Time).notNullable().primary();
+      table.bigInteger(SpeedTimesColumn.Time).notNullable().primary();
       table.integer(SpeedTimesColumn.Speed).nullable();
     });
   }
